requirements.txt
fastapi
uvicorn
requests
tensorflow
numpy
pandas
scipy
scikit-learn
Django
djangorestframework
gunicorn
